domain_generalization/image_prediction_service/requirements.txt
flask
pillow
numpy
gunicorn
flask_cors
torch
torchvision
matplotlib